+++ packages/create-jsenv/demo-web-preact/scripts/test.mjs
@@ -8,14 +8,16 @@ import { pingServer, executeTestPlan, chromium, webkit } from "@jsenv/core"
 
 import { rootDirectoryUrl, plugins } from "../jsenv.config.mjs"
 
-const devServerStarted = await pingServer(`http://127.0.0.1:3400`)
+const devServerOrigin = "http://localhost:3401"
+const devServerStarted = await pingServer(devServerOrigin)
 let devServer
 if (!devServerStarted) {
-  devServer = await import("./start_dev_server.mjs").devServer
+  devServer = (await import("./start_dev_server.mjs")).devServer
 }
 try {
   await executeTestPlan({
     rootDirectoryUrl,
+    devServerOrigin,
     plugins,
     testPlan: {
       "./tests/**/*.test.html": {